Created some new classes to clean up the level.cpp code.
Created enemyWave and updated killer and player classes to include
collision detection. May move collision detection to it's own class, but
it's fairly simple as it is. Also updated the sound assets w/ new sound
for the boss (killer). Other minor tweaks to the ammo and enemy classes
were made.